requirements.txt
psycopg2-binary 
python-dotenv
bcrypt
fastapi 
uvicorn
sqlalchemy
python-jose[cryptography]